#14 feat: 가독성 개선을 위한 패키지 분류
