+++ backend/src/main/java/moheng/auth/exception/InvalidTokenFormatException.java
@@ -1,4 +1,4 @@
-package moheng.auth.presentation;
+package moheng.auth.exception;
 
 public class InvalidTokenFormatException extends RuntimeException {
     public InvalidTokenFormatException(String message) {
