+++ backend/src/main/java/moheng/auth/presentation/admin/AdminAuthentication.java
@@ -1,4 +1,4 @@
-package moheng.auth.presentation;
+package moheng.auth.presentation.admin;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
