+++ backend/src/main/java/moheng/auth/presentation/admin/AdminAuthenticationArgumentResolver.java
@@ -1,8 +1,9 @@
-package moheng.auth.presentation;
+package moheng.auth.presentation.admin;
 
 import jakarta.servlet.http.HttpServletRequest;
 import moheng.auth.domain.JwtTokenProvider;
 import moheng.auth.dto.LoginMember;
+import moheng.auth.presentation.authentication.AuthenticationBearerExtractor;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -11,11 +12,11 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
 @Component
-public class AuthenticationArgumentResolver implements HandlerMethodArgumentResolver {
+public class AdminAuthenticationArgumentResolver implements HandlerMethodArgumentResolver {
     private final JwtTokenProvider jwtTokenProvider;
     private final AuthenticationBearerExtractor authenticationBearerExtractor;
 
-    public AuthenticationArgumentResolver(final JwtTokenProvider jwtTokenProvider, final AuthenticationBearerExtractor authenticationBearerExtractor) {
+    public AdminAuthenticationArgumentResolver(final JwtTokenProvider jwtTokenProvider, final AuthenticationBearerExtractor authenticationBearerExtractor) {
         this.jwtTokenProvider = jwtTokenProvider;
         this.authenticationBearerExtractor = authenticationBearerExtractor;
     }
@@ -32,6 +33,6 @@ public class AuthenticationArgumentResolver implements HandlerMethodArgumentReso
 
     @Override
     public boolean supportsParameter(final MethodParameter methodParameter) {
-        return methodParameter.hasMethodAnnotation(Authentication.class);
+        return methodParameter.hasMethodAnnotation(AdminAuthentication.class);
     }
 }
