+++ backend/src/main/java/moheng/auth/presentation/authentication/AuthenticationBearerExtractor.java
@@ -1,7 +1,8 @@
-package moheng.auth.presentation;
+package moheng.auth.presentation.authentication;
 
 import jakarta.servlet.http.HttpServletRequest;
 import moheng.auth.exception.EmptyBearerHeaderException;
+import moheng.auth.exception.InvalidTokenFormatException;
 import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Component;
