+++ backend/src/main/java/moheng/auth/presentation/authentication/AuthenticationWebConfig.java
@@ -1,6 +1,5 @@
-package moheng.auth.presentation;
+package moheng.auth.presentation.authentication;
 
-import org.springframework.boot.actuate.autoconfigure.metrics.MetricsProperties;
 import org.springframework.stereotype.Component;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
